Support for authentication via tokens
For the first login, email will be used to request an authentication token.
It will then be stored in a JSON file (you can create the JSON file in
advance with appropriate rights if you require more security).

As for now, I don't know how token expiration should be handled, more commits
will probably come to handle this case.